// frontend/src/api/openapi/models/Player.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Player = {
    id: number;
    fantasyId: number;
    firstName: string;
    lastName: string;
    positionId: number;
    teamId: number;
    positionCategory: string;
    cost: number;
    displayName: string;
    imagePath: string;
    nationality: string;
};