// frontend/src/api/openapi/models/FootballTeam.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FootballTeam = {
    id: number;
    name: string;
    shortCode: string;
    logoPath: string;
    venueId: number;
    legacyId: number;
    leagueId: number;
};