// frontend/src/api/openapi/models/SeasonPlayerStats.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Player } from './Player';
import type { PlayerStats } from './PlayerStats';
export type SeasonPlayerStats = {
    stats: PlayerStats;
    player: Player;
    gamesPlayed?: number;
};